Tentando Salvar foto

+++ src/main/java/com/ufcg/backend/services/ReservaService.java
@@ -56,7 +56,9 @@ public class ReservaService {
     }
 
     public void deletarReservasByIdArea(Long id) {
-        reservaRepository.deleteAllByIdArea(id);
+        if(!reservaRepository.getAllReservaByIdArea(id).isEmpty()) {
+            reservaRepository.deleteAllByIdArea(id);
+        }
     }
 
     public void deletarReservasByIdUser(Long id) {
